[📦 NEW] SR_EV_UPDATE

+++ main.cpp
@@ -17,6 +17,9 @@ struct SysrepoCallback : Callback {
     // Display event type.
     std::ostringstream event_type;
     switch (event) {
+    case SR_EV_UPDATE:
+      event_type << "update";
+      break;
     case SR_EV_CHANGE:
       event_type << "change";
       break;
@@ -60,19 +63,16 @@ struct SysrepoCallback : Callback {
       CHANGES.push_back(change);
     }
 
-    // Get an item.
-    S_Val value(session->get_item("/model:config/child"));
-
-    std::thread([]() {
-      S_Connection connection(std::make_shared<Connection>());
-      S_Session session(std::make_shared<Session>(connection, SR_DS_RUNNING));
-      session->delete_item("/model:config");
-      session->set_item("/model:config/le_list[name='whatever']/contained/data",
-                        std::make_shared<Val>(1337));
-      session->set_item_str(
-          "/model:config/le_list[name='whatever']/contained/floating", "0.37");
-      session->apply_changes();
-    }).detach();
+    if (event == SR_EV_UPDATE) {
+        S_Val value(session->get_item("/model:config/child"));
+        session->delete_item("/model:config");
+        session->set_item("/model:config/le_list[name='whatever']/contained/data",
+            std::make_shared<Val>(1337));
+        session->set_item_str(
+            "/model:config/le_list[name='whatever']/contained/floating", "0.37");
+        session->set_item("/model:config/child",
+            std::make_shared<Val>(1337));
+    }
 
     return 0;
   }
@@ -85,7 +85,7 @@ struct SysrepoClient {
     subscription_ = std::make_shared<Subscribe>(session_);
     subscription_->module_change_subscribe(model_.c_str(),
                                            std::make_shared<SysrepoCallback>(),
-                                           0, 0, SR_SUBSCR_DEFAULT);
+                                           0, 0, SR_SUBSCR_UPDATE);
   }
 
   void displayChanges() {
